drivers/i2c/i2c-aspeed: Only use IRQ status for command completion notification
We currently determine if an i2c command as complete by looking at the command
register (and comparing it with the pending set of commands). While this
may indicate that all pending commands have completed, we may still get
a further interrupt when multiple commands (like a read + stop) have
been queued.

This change uses the interrupt status to indicate command completion,
instead of the command register.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>